feat: add soft delete feature

+++ packages/slonik/src/sqlFactory.ts
@@ -37,6 +37,7 @@ class DefaultSqlFactory implements SqlFactory {
   protected _factory: SqlFactory | undefined;
   protected _schema = "public";
   protected _validationSchema: z.ZodTypeAny = z.any();
+  protected _softDeleteEnabled: boolean = false;
 
   constructor(config: ApiConfig, database: Database, schema?: string) {
     this._config = config;
@@ -66,6 +67,7 @@ class DefaultSqlFactory implements SqlFactory {
     return sql.type(allSchema)`
       SELECT ${sql.join(identifiers, sql.fragment`, `)}
       FROM ${this.getTableFragment()}
+      ${this.getSoftDeleteFilterFragment(true)}
       ${this.getSortFragment(sort)}
     `;
   }
@@ -78,7 +80,8 @@ class DefaultSqlFactory implements SqlFactory {
     return sql.type(countSchema)`
       SELECT COUNT(*)
       FROM ${this.getTableFragment()}
-      ${this.getFilterFragment(filters)};
+      ${this.getFilterFragment(filters)}
+      ${this.getSoftDeleteFilterFragment(!filters)};
     `;
   }
 
@@ -105,7 +108,16 @@ class DefaultSqlFactory implements SqlFactory {
     `;
   }
 
-  getDeleteSql(id: number | string): QuerySqlToken {
+  getDeleteSql(id: number | string, force: false): QuerySqlToken {
+    if (this.softDeleteEnabled && !force) {
+      return sql.type(this.validationSchema)`
+        UPDATE ${this.getTableFragment()}
+        SET ${this.tableIdentifier}.deleted_at = NOW()
+        WHERE id = ${id}
+        RETURNING *;
+      `;
+    }
+
     return sql.type(this.validationSchema)`
       DELETE FROM ${this.getTableFragment()}
       WHERE id = ${id}
@@ -117,7 +129,8 @@ class DefaultSqlFactory implements SqlFactory {
     return sql.type(this.validationSchema)`
       SELECT *
       FROM ${this.getTableFragment()}
-      WHERE id = ${id};
+      WHERE id = ${id}
+      ${this.getSoftDeleteFilterFragment(false)};
     `;
   }
 
@@ -126,6 +139,7 @@ class DefaultSqlFactory implements SqlFactory {
       SELECT *
       FROM ${this.getTableFragment()}
       ${this.getFilterFragment(filters)}
+      ${this.getSoftDeleteFilterFragment(!filters)}
       ${this.getSortFragment(sort)}
       LIMIT 1;
     `;
@@ -136,6 +150,7 @@ class DefaultSqlFactory implements SqlFactory {
       SELECT *
       FROM ${this.getTableFragment()}
       ${this.getFilterFragment(filters)}
+      ${this.getSoftDeleteFilterFragment(!filters)}
       ${this.getSortFragment(sort)};
     `;
   }
@@ -150,6 +165,7 @@ class DefaultSqlFactory implements SqlFactory {
       SELECT *
       FROM ${this.getTableFragment()}
       ${this.getFilterFragment(filters)}
+      ${this.getSoftDeleteFilterFragment(!filters)}
       ${this.getSortFragment(sort)}
       ${this.getLimitFragment(limit, offset)};
     `;
@@ -181,6 +197,7 @@ class DefaultSqlFactory implements SqlFactory {
       UPDATE ${this.getTableFragment()}
       SET ${sql.join(columns, sql.fragment`, `)}
       WHERE id = ${id}
+      ${this.getSoftDeleteFilterFragment(false)}
       RETURNING *;
     `;
   }
@@ -231,6 +248,10 @@ class DefaultSqlFactory implements SqlFactory {
     return this._validationSchema || z.any();
   }
 
+  get softDeleteEnabled(): boolean {
+    return this._softDeleteEnabled;
+  }
+
   protected getFilterFragment(filters?: FilterInput): FragmentSqlToken {
     return createFilterFragment(filters, this.tableIdentifier);
   }
@@ -242,6 +263,14 @@ class DefaultSqlFactory implements SqlFactory {
     limit = Math.min(limit ?? this.limitDefault, this.limitMax);
 
     return createLimitFragment(limit, offset);
+  }
+
+  protected getSoftDeleteFilterFragment(addWhere: boolean): FragmentSqlToken {
+    return this._softDeleteEnabled
+      ? addWhere
+        ? sql.fragment`WHERE ${this.tableIdentifier}.deleted_at IS NULL`
+        : sql.fragment`AND ${this.tableIdentifier}.deleted_at IS NULL`
+      : sql.fragment``;
   }
 
   protected getSortFragment(sort?: SortInput[]): FragmentSqlToken {
